affile: make PollFileChanges inheritable
